gibbs3/library: added interp_SA_CT and interp_ref_cast; symlink for data
It looks like the original gsw_* versions are just doing simple
linear interpolation, so I used the much simpler linear interpolation
scheme from scipy.interpolate to make _interp_S_T, and then used
that for both of the gsw_* functions.

We may not want the data symlink long-term, but it works for now
for testing, at least on *nix.